fix set loading issue
